v0.1

+++ src/seeders/20210520032847-init.ts
@@ -2,9 +2,10 @@
 import { QueryInterface } from "sequelize";
 import { getListPermissions, PermissionsList, getListRolesPermissions } from "../constants";
 import { encrypt } from "../utils/security";
+import { adminObject } from "../utils/types"
+import config from 'config';
 
-
-const email = "byruk228i@gmail.com";
+const admin: adminObject = config.get("admin");
 
 module.exports = {
     up: async (queryInterface: QueryInterface) => {
@@ -22,10 +23,10 @@ module.exports = {
         await queryInterface.bulkInsert('permissions', getListPermissions(PermissionsList))
         const permissions: any = (await queryInterface.sequelize.query('SELECT * FROM "permissions"'))[0];
         const roles: any = (await queryInterface.sequelize.query('SELECT * FROM "roles"'))[0];
-        const encryptData = await encrypt('test');
+        const encryptData = await encrypt(admin.password);
         await queryInterface.bulkInsert('role_permissions', getListRolesPermissions(PermissionsList, permissions, roles))
         await queryInterface.insert(null, "users", {
-            email,
+            email: admin.email,
             password: encryptData.encryptedPassword,
             password_salt: encryptData.salt,
             roleId: roles[roles.findIndex((item: any) => item.name === 'ADMIN')].id
@@ -33,7 +34,7 @@ module.exports = {
     },
 
     down: async (queryInterface: QueryInterface) => {
-        await queryInterface.sequelize.query(`DELETE FROM "users" WHERE email='${ email }'`);
+        await queryInterface.sequelize.query(`DELETE FROM "users" WHERE email='${ admin.email }'`);
         await queryInterface.sequelize.query('DELETE FROM "role_permissions"');
         await queryInterface.sequelize.query('DELETE FROM "roles"');
         await queryInterface.sequelize.query('DELETE FROM "permissions"');
